RectInt使うのやめる

+++ PsdPlugin/PsdPlugin/PsdFile/Compression/EndianReverser.cs
@@ -34,7 +34,7 @@ namespace PhotoshopFile.Compression
         {
             imageData.Read(buffer);
 
-            var numPixels = Size.x * Size.y;
+            var numPixels = (int)Size.x * (int)Size.y;
             if (numPixels == 0)
             {
                 return;
@@ -52,7 +52,7 @@ namespace PhotoshopFile.Compression
             // Reverse endianness before passing on to underlying compressor
             if (array.Length > 0)
             {
-                var numPixels = array.Length / BytesPerRow * Size.x;
+                var numPixels = array.Length / BytesPerRow * (int)Size.x;
                 Util.SwapByteArray(BitDepth, array, 0, numPixels);
             }
